:heavy_plus_sign: Add HomeScreen & Code Improve

+++ app/src/main/java/com/ashish/nowplayingapp/model/FavMovie.kt
@@ -6,5 +6,5 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "fav_movies")
 data class FavMovie(
     @PrimaryKey
-    val id : Int
+    val id : Long
 )
